gpu: nvgpu: fix K4.9 sparse messages
Fix issues related with wrong storage type for
64 bit variables.

(1) Fixed width of HZ_TO_MHZ constant
(2) changed fence_wait timeout to store unsigned
long

bug 200299572

Reviewed-on: http://git-master/r/1471611

+++ drivers/gpu/nvgpu/gk20a/fence_gk20a.c
@@ -82,11 +82,12 @@ static inline bool gk20a_fence_is_valid(struct gk20a_fence *f)
 	return valid;
 }
 
-int gk20a_fence_wait(struct gk20a *g, struct gk20a_fence *f, int timeout)
+int gk20a_fence_wait(struct gk20a *g, struct gk20a_fence *f,
+							unsigned long timeout)
 {
 	if (f && gk20a_fence_is_valid(f)) {
 		if (!nvgpu_platform_is_silicon(g))
-			timeout = (u32)MAX_SCHEDULE_TIMEOUT;
+			timeout = MAX_SCHEDULE_TIMEOUT;
 		return f->ops->wait(f, timeout);
 	}
 	return 0;
